requirements.txt
adbutils~=2.10.2; python_version >= "3.10" and python_version < "3.14"
annotated-types==0.7.0; python_version >= "3.10" and python_version < "3.14"
anyio==4.10.0; python_version >= "3.10" and python_version < "3.14"
click==8.2.1; python_version >= "3.10" and python_version < "3.14"
colorama==0.4.6; python_version >= "3.10" and python_version < "3.14"
easyocr==1.7.2; python_version >= "3.10" and python_version < "3.14"
exceptiongroup==1.3.0; python_version >= "3.10" and python_version < "3.14"
fastapi==0.116.1; python_version >= "3.10" and python_version < "3.14"
filelock==3.18.0; python_version >= "3.10" and python_version < "3.14"
fsspec==2025.7.0; python_version >= "3.10" and python_version < "3.14"
h11==0.16.0; python_version >= "3.10" and python_version < "3.14"
idna==3.10; python_version >= "3.10" and python_version < "3.14"
imageio==2.37.0; python_version >= "3.10" and python_version < "3.14"
Jinja2==3.1.6; python_version >= "3.10" and python_version < "3.14"
keyboard==0.13.5; python_version >= "3.10" and python_version < "3.14"
lazy_loader==0.4; python_version >= "3.10" and python_version < "3.14"
Levenshtein==0.27.1; python_version >= "3.10" and python_version < "3.14"
MarkupSafe==3.0.2; python_version >= "3.10" and python_version < "3.14"
MouseInfo==0.1.3; python_version >= "3.10" and python_version < "3.14"
mpmath==1.3.0; python_version >= "3.10" and python_version < "3.14"
mss==10.0.0; python_version >= "3.10" and python_version < "3.14"
networkx==3.4.2; python_version >= "3.10" and python_version < "3.14"
ninja==1.11.1.4; python_version >= "3.10" and python_version < "3.14"
numpy==2.2.6; python_version >= "3.10" and python_version < "3.14"
opencv-python==4.12.0.88; python_version >= "3.10" and python_version < "3.14"
opencv-python-headless==4.12.0.88; python_version >= "3.10" and python_version < "3.14"
packaging==25.0; python_version >= "3.10" and python_version < "3.14"
pillow==11.3.0; python_version >= "3.10" and python_version < "3.14"
PyAutoGUI==0.9.54; python_version >= "3.10" and python_version < "3.14"
pyclipper==1.3.0.post6; python_version >= "3.10" and python_version < "3.14"
pyobjc-framework-Vision==12.0; python_version >= "3.10" and python_version < "3.14" and sys_platform == "darwin"
pydantic==2.11.7; python_version >= "3.10" and python_version < "3.14"
pydantic_core==2.33.2; python_version >= "3.10" and python_version < "3.14"
PyGetWindow==0.0.9; python_version >= "3.10" and python_version < "3.14"
PyMsgBox==1.0.9; python_version >= "3.10" and python_version < "3.14"
pyperclip==1.9.0; python_version >= "3.10" and python_version < "3.14"
PyRect==0.2.0; python_version >= "3.10" and python_version < "3.14"
PyScreeze==1.0.1; python_version >= "3.10" and python_version < "3.14"
python-bidi==0.6.6; python_version >= "3.10" and python_version < "3.14"
pytweening==1.2.0; python_version >= "3.10" and python_version < "3.14"
PyYAML==6.0.2; python_version >= "3.10" and python_version < "3.14"
RapidFuzz==3.13.0; python_version >= "3.10" and python_version < "3.14"
scikit-image==0.25.2; python_version >= "3.10" and python_version < "3.14"
scipy==1.15.3; python_version >= "3.10" and python_version < "3.14"
shapely==2.1.1; python_version >= "3.10" and python_version < "3.14"
sniffio==1.3.1; python_version >= "3.10" and python_version < "3.14"
starlette==0.47.2; python_version >= "3.10" and python_version < "3.14"
sympy==1.14.0; python_version >= "3.10" and python_version < "3.14"
tifffile==2025.5.10; python_version >= "3.10" and python_version < "3.14"
torch==2.7.1; python_version >= "3.10" and python_version < "3.14"
torchaudio==2.7.1; python_version >= "3.10" and python_version < "3.14"
torchvision==0.22.1; python_version >= "3.10" and python_version < "3.14"
typing-inspection==0.4.1; python_version >= "3.10" and python_version < "3.14"
typing_extensions==4.14.1; python_version >= "3.10" and python_version < "3.14"
uvicorn==0.35.0; python_version >= "3.10" and python_version < "3.14"
